Add new_shape('chain'...), switch to mediump, add screen2scene (zoom unhandled)

+++ src/engine.cpp
@@ -102,7 +102,7 @@ void Engine::update()
 
 	AT(event)
 
-	double dt = (get_now() - last_update) / 1000;
+	float dt = (get_now() - last_update) / 1000;
 	last_update = get_now();
 
 	if (update_activated)
